# web_api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Reabpc api",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/build/index": {
            "get": {
                "tags": [
                    "guest",
                    "user"
                ],
                "summary": "index pagination components price by shopID",
                "operationId": "App\\Http\\Controllers\\BuildpcController::index",
                "parameters": [
                    {
                        "name": "adminshopID",
                        "in": "query",
                        "description": "id of component",
                        "required": true,
                        "example": "1"
                    },
                    {
                        "name": "component",
                        "in": "query",
                        "description": "name of component",
                        "required": true,
                        "example": "cpu"
                    },
                    {
                        "name": "current_page",
                        "in": "query",
                        "description": "number of component",
                        "required": false,
                        "example": "10"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "number of pagination",
                        "required": false,
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "please test it."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/build/list": {
            "get": {
                "tags": [
                    "guest",
                    "user"
                ],
                "summary": "list components price by shopID",
                "parameters": [
                    {
                        "name": "adminshopID",
                        "in": "query",
                        "description": "id of component",
                        "required": true,
                        "example": "1"
                    },
                    {
                        "name": "component",
                        "in": "query",
                        "description": "name of component",
                        "required": true,
                        "example": "cpu"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "please test it."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/admin_shop/components/index": {
            "get": {
                "tags": [
                    "shop"
                ],
                "summary": "index pagination components",
                "operationId": "App\\Http\\Controllers\\ComponentController::index",
                "parameters": [
                    {
                        "name": "component",
                        "in": "query",
                        "description": "name of component",
                        "required": true,
                        "example": "cpu"
                    },
                    {
                        "name": "current_page",
                        "in": "query",
                        "description": "number of component",
                        "required": false,
                        "example": "10"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "number of pagination",
                        "required": false,
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "please test it."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/admin_shop/components/list": {
            "get": {
                "tags": [
                    "shop"
                ],
                "summary": "list components",
                "parameters": [
                    {
                        "name": "component",
                        "in": "query",
                        "description": "name of component",
                        "required": true,
                        "example": "cpu"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "please test it."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/admin_shop/profile_info": {
            "get": {
                "tags": [
                    "shop"
                ],
                "summary": "shop profile info",
                "operationId": "App\\Http\\Controllers\\ShopController::profileInfo",
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "please test it."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/profile_info": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "user profile info",
                "operationId": "App\\Http\\Controllers\\UserController::profileInfo",
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "please test it."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        []
    ]
}